requirements.txt
aniso8601==9.0.1
APScheduler==3.10.4
attrs==23.1.0
beautifulsoup4==4.12.3
bidict==0.22.1
blinker==1.6.2
bs4==0.0.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.3
colorama==0.4.6
contourpy==1.2.0
cryptography==40.0.2
cycler==0.12.1
decorator==5.1.1
dnspython==2.4.1
et-xmlfile==1.1.0
eventlet==0.33.3
Flask==2.3.1
Flask-Login==0.6.2
flask-restx==1.1.0
Flask-SocketIO==5.3.5
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.1.1
fonttools==4.47.2
greenlet==2.0.2
idna==3.6
importlib-metadata==7.0.0
importlib-resources==6.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
jsonify==0.5
jsonpatch==1.33
jsonpath-rw==1.4.0
jsonpointer==2.4
jsonschema==4.17.3
kiwisolver==1.4.5
ldap3==2.9.1
MarkupSafe==2.1.2
matplotlib==3.8.2
nltk==3.8.1
numpy==1.26.2
openpyxl==3.1.2
packaging==23.2
pandas==2.1.3
pillow==10.2.0
pkgutil_resolve_name==1.3.10
ply==3.11
psycopg2==2.9.6
psycopg2-binary==2.9.9
pyasn1==0.5.0
pycparser==2.21
pycryptodome==3.17
pyparsing==3.1.1
pyrsistent==0.19.3
pyspnego==0.10.2
python-dateutil==2.8.2
python-engineio==4.5.1
python-socketio==5.8.0
pytz==2023.3
redfish==3.2.2
regex==2023.12.25
requests==2.31.0
requests-kerberos==0.14.0
requests-toolbelt==1.0.0
requests-unixsocket==0.3.0
rsa==4.9
six==1.16.0
soupsieve==2.5
SQLAlchemy==2.0.25
sspilib==0.1.0
touch==2020.12.3
tqdm==4.66.1
typing_extensions==4.9.0
tzdata==2023.3
tzlocal==5.2
urllib3==2.1.0
values==2020.12.3
Werkzeug==2.3.0
WTForms==3.0.1
zipp==3.17.0
